Cleanup. If something doesn't have Override annotation, check it.

+++ src/main/java/mods/natura/blocks/GrassSlab.java
@@ -91,14 +91,14 @@ public class GrassSlab extends NSlabBase
                 int grassColor = 0;
                 if (meta == 1)
                 {
-                    double temp = (double) MathHelper.clamp_float(biome.getFloatTemperature(x, y, z), 0.0F, 1.0F);
-                    double rainfall = (double) MathHelper.clamp_float(biome.getFloatRainfall(), 0.0F, 1.0F);
+                    double temp = MathHelper.clamp_float(biome.getFloatTemperature(x, y, z), 0.0F, 1.0F);
+                    double rainfall = MathHelper.clamp_float(biome.getFloatRainfall(), 0.0F, 1.0F);
                     grassColor = GrassColorizerAlternate.getBlueGrassColor(temp, rainfall);
                 }
                 else if (meta == 2)
                 {
-                    double temp = (double) MathHelper.clamp_float(biome.getFloatTemperature(x, y, z), 0.0F, 1.0F);
-                    double rainfall = (double) MathHelper.clamp_float(biome.getFloatRainfall(), 0.0F, 1.0F);
+                    double temp = MathHelper.clamp_float(biome.getFloatTemperature(x, y, z), 0.0F, 1.0F);
+                    double rainfall = MathHelper.clamp_float(biome.getFloatRainfall(), 0.0F, 1.0F);
                     grassColor = GrassColorizerAlternate.getOrangeGrassColor(temp, rainfall);
                 }
                 else
